tests/wpt/tests/css/css-backgrounds/parsing/background-attachment-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Backgrounds and Borders Module Level 3: parsing background-attachment with invalid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-backgrounds/#background-attachment">
<meta name="assert" content="background-attachment supports only the grammar '<attachment>#'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("background-attachment", "auto");
test_invalid_value("background-attachment", "local, none");
</script>
</body>
</html>
